Use the system confget instead of our own.
Let the storpool-config-dep package depend on "confget": on recent
versions of Debian and Ubuntu it is available in the system
repositories, for older versions of Ubuntu and also all versions of
CentOS we have packaged it in the storpool-contrib repository.

Remove the /usr/lib/storpool/confget file and replace it with a symlink
to /usr/bin/confget for the benefit of external tools that might
reference it for some weird reason (e.g. "-q sections" until recently).

Adapt all the tools that invoked confget directly to use either
storpool_confget (mainly for "-q sections") or storpool_confshow
(mainly for "-n -e -s otherhostname VARNAME").

Drop the Makefile in the contrib/ directory since it does not build or
install anything now, and do not recurse into the contrib/ directory
from the top-level Makefile.

Drop the mention of confget from the license list files.

Change-Id: I48604c0b658f5c5c94e389c065b1fb1aef99926f