mgmt/bridge/block/cli/bindings/iscsi: implement revert to snapshot
Issue is mainly with volumes with no name - we must create a new volume and it will have a new globalId, so without the Revert API call database update for the VM managment system will be needed.

The actual implementation of the rever call is an absolute unit of a hack. We store the ~old.global.id as the name of the new volume and when searching for volume by globalId we also check for names if one is not found by globalId

The spbd part is to create a symlink /dev/storpool-byid/old.global.id instead of /dev/storpool/~old.global.id

For the iscsi use the preserved visible volume id as a serial number, so that the reverted volume will look the same for the user

Change-Id: I6168d6592e9376817e28c52b38383a11226ea402