python-api: error out on missing config files.
Expect /usr/lib/storpool/storpool-defaults.conf and /etc/storpool.conf
to be present. Also, expect anything named *.conf found in
/etc/storpool.conf.d/ to be readable and parseable.

The StorPool installation procedures now always include a "place this
/etc/storpool.conf file on the server" before the actual installation,
so it is time to flip this switch.

If the "missing_ok=True" named parameter is passed to SPConfig() or to
SPConfig.get_config_files(), there is no change in behavior. This may be
used in tools that need to run really, really early during the StorPool
installation, or in tools that need to gather information and may have
to handle "incomplete StorPool configuration" as a failure mode.

Change-Id: I5605884b2eeeaeee91060b8255a093494dd55749