python-api: hardcode the default settings.
Do not even attempt to read the /usr/lib/storpool/storpool-defaults.conf
file, but keep its contents as of the time of this commit as static
initialization values. This, along with the missing_ok parameter, will
make it easier to use an updated version of SPConfig even before
the full installation of StorPool has been completed.

Change-Id: Iffc9d1d092aa551294246a18c47cfe333522f698