python-api: return the raw API response if requested.
Add a boolean returnRawAPIData parameter to each API call to make it
return the data immediately after parsing the JSON response and retrying
on transient errors, before validating it as the return type class and
possibly cutting off fields that the Python bindings do not yet know
about.

Please only use this if absolutely needed and... very, very carefully.
Please consider asking for the fields you need to be added to the Python
bindings before using this mechanism.

Change-Id: I4c9e86ee418737cc714337975e103a21e3498aa1
ObQuote: "This place is not a place of honor... no highly esteemed deed is commemorated here.. nothing valued is here."