Add storpool-logd
Add initial files to the "storpool-logd" project.

This project aims to implement a log daemon. It will read log
messages and kernel messages (each from its own source) and send them to
an HTTPS endpoint.

All information is kept in the README.md file

This logd is also added as a service through the storpool-daemon.

Change-Id: I64ad3920ffce35a5379964c4b51b306f4d19ece3